+++ functions/index.js
@@ -15,3 +15,4 @@ exports.onContratoOrdenWrite          = onContratoOrdenWrite;
 exports.onOrdenWriteSyncContratoCache = onOrdenWriteSyncContratoCache;
 exports.onOrdenHardDelete             = onOrdenHardDelete;
 exports.onOrdenWriteSearchTokens      = require("./src/triggers/ordenes/onWriteSearchTokens");
+exports.purgePIIRetention             = require("./src/triggers/scheduled/purgePIIRetention");